requirements.txt
altgraph==0.17
docopt==0.6.2
importlib-metadata==4.3.0
macholib==1.14
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
typing-extensions==3.10.0.0
webvtt-py==0.4.6
youtube-dl==2021.5.16
zipp==3.4.1
